requirements.txt
gradio
openai
python-dotenv
requests
wikipedia
html-to-markdown
beautifulsoup4
langchain
langchain-core
langchain-tavily